Login added

+++ app.py
@@ -1,6 +1,7 @@
 import sqlite3
-from flask import Flask, render_template, request, url_for, flash, redirect
+from flask import Flask, render_template, request, url_for, flash, redirect,session, abort
 from werkzeug.exceptions import abort
+import os
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'your secret key'
 
@@ -88,31 +89,37 @@ def get_student(roll_no):
     conn.close()
     return student
 
-@app.route('/students/edit/<string:roll_no>',methods = ('GET','POST'))
 def editstudent(roll_no):
-    student = get_student(roll_no)
+    if not session.get('logged_in'):
+        return render_template('login.html')
+    else:
+        student = get_student(roll_no)
 
-    if request.method == 'POST':
-        #roll_no = request.form['roll_no']
-        name = request.form['name']
-        email = request.form['email']
-        committee = request.form['committee']
-        year = request.form['year']
-        phone = request.form['phone']
+        if request.method == 'POST':
+            #roll_no = request.form['roll_no']
+            name = request.form['name']
+            email = request.form['email']
+            committee = request.form['committee']
+            year = request.form['year']
+            phone = request.form['phone']
 
-        if not name:
-            flash('Name is required!')
-        else:
-            conn = get_db_connection()
-            conn.execute('UPDATE Student SET Name = ?'
-                         ', Email = ?, Committee = ?, Year = ?, Phone = ? WHERE Roll_no = ?',
-                         (name,email,committee,year,phone,roll_no))
-            conn.commit()
-            conn.close()
-            return redirect(url_for('allstudents'))
+            if not name:
+                flash('Name is required!')
+            else:
+                conn = get_db_connection()
+                conn.execute('UPDATE Student SET Name = ?'
+                             ', Email = ?, Committee = ?, Year = ?, Phone = ? WHERE Roll_no = ?',
+                             (name,email,committee,year,phone,roll_no))
+                conn.commit()
+                conn.close()
+                return redirect(url_for('index'))
+
+        return render_template('editstudent.html',student = student )
     
 @app.route('/events/<string:event_id>/remove/workson/<string:roll_no>',methods = ('POST',))
 def remove_works_on(roll_no,event_id):
+    if not session.get('logged_in'):
+        return render_template('login.html')
     conn = get_db_connection()
     conn.execute('DELETE FROM Works_on WHERE Roll_no = ? AND Event_ID = ?', (roll_no,event_id))
     conn.commit()
@@ -139,6 +146,8 @@ def remove_winner(event_id,pid):
 
 @app.route('/students/newstudent',methods = ('GET','POST'))
 def newstudent():
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
             name = request.form['name']
             roll_no = request.form['roll_no']
@@ -159,6 +168,8 @@ def newstudent():
 
 @app.route('/events/newevent', methods = ('GET','POST'))
 def newevent():
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
             event_name = request.form['event_name']
             event_id = request.form['event_id']
@@ -177,6 +188,8 @@ def newevent():
 
 @app.route('/judges/newjudge',methods = ('GET','POST'))
 def newjudge():
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
             name = request.form['name']
             judge_id = request.form['judge_id']
@@ -194,6 +207,8 @@ def newjudge():
 
 @app.route('/participants/newparticipant',methods = ('GET','POST'))
 def newparticipant():
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
             name = request.form['name']
             pid = request.form['pid']
@@ -212,6 +227,8 @@ def newparticipant():
 
 @app.route('/venues/newvenue',methods = ('GET','POST'))
 def newvenue():
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
             venue_name = request.form['venue_name']
             venue_id = request.form['venue_id']
@@ -248,6 +265,8 @@ def newwinner(event_id):
 
 @app.route('/events/<string:event_id>/newrequirement', methods = ('GET','POST'))
 def newrequirement(event_id):
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
         requirement = request.form['requirement']
         count = request.form['count']
@@ -278,6 +297,8 @@ def newparticipation(event_id):
 
 @app.route('/events/<string:event_id>/newwork', methods = ('GET','POST'))
 def newwork(event_id):
+    if not session.get('logged_in'):
+        return render_template('login.html')
     if request.method == 'POST':
         roll_no = request.form['roll_no']
         #reception_status = request.form['reception_status']
@@ -361,3 +382,17 @@ def alljudges():
 @app.route("/")
 def  index():
     return render_template('indexf2.html')
+    
+@app.route("/logout")
+def logout():
+    session['logged_in'] = False
+    return  redirect(url_for('index'))
+
+@app.route('/login', methods=['POST'])
+def do_admin_login():
+    if request.form['password'] == 'password' and request.form['username'] == 'admin':
+        session['logged_in'] = True
+        return redirect(url_for('index'))
+    else:
+        flash('wrong password!')
+        return redirect(url_for('index'))
